Disabled writing of hbm files because runs are to be on emft server

+++ tests/org.eclipse.emf.teneo.hibernate.test/src/org/eclipse/emf/teneo/hibernate/test/HibernateTestbed.java
@@ -37,7 +37,7 @@ import org.eclipse.emf.teneo.test.stores.TestStore;
  * Is the testbed which models the base in which a testrun is run.
  * 
  * @author <a href="mailto:mtaal@elver.org">Martin Taal</a>
- * @version $Revision: 1.1 $
+ * @version $Revision: 1.2 $
  */
 public class HibernateTestbed extends Testbed {
 
@@ -104,6 +104,10 @@ public class HibernateTestbed extends Testbed {
 
 	/** Write the mapping file for debugging purposes */
 	private void writeMappingToFile(File mappingFile, String mappingXML) throws IOException {
+		if (true) {
+			log.debug("Disabled writing of hibernate.hbm.xml file for tests on the vserver");
+			return;
+		}
 		if (mappingFile.exists()) mappingFile.delete();
 		mappingFile.createNewFile();
 		final FileWriter fileWriter = new FileWriter(mappingFile);
